services/dashboard/requirements.txt
# Dashboard Service Dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.1
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
jinja2==3.1.2
loguru==0.7.0
email-validator==2.0.0
prometheus-client==0.16.0
redis==4.5.5